test(ApiStructureTest): reenable test for visualeditoredit integration
The test was double broken:
First because ApiStructureTest reused the
ApiMain instance which caused the ObjectFactory injected into
GrowthExperimentsTaskTypeHandlerRegistry to reference a destroyed
service container. See I3af04966b for more details.

And second because ApiParamValidator did not yet know that it is also
valid to use ApiBase::PARAM_HELP_MSG_PER_VALUE to document parameters
with multi-string type. See Ia8c1abdcb6 for that.

Bug: T380585
Depends-On: Ia8c1abdcb68823650582208caa9be3fa2c47daa8
Change-Id: I1167bb8ad3721d46a0004f0c79a148adbc0f2633