test(Cypress): Introduce GuidedTour page-object
This should make it more readable to get these popups out of the way.

Also, it does not make sense to try to disable the homepage_discovery
popup via setOptions because it does that by itself when it shows on the
first pageview after log-in. It generally does make sense to disable the
homepage_welcome popup that way.

However, the homepage_welcome popup is explicitly tested in the homepage
browser test, because that seems to be the right place to test that.

Follow-Up-To: Ied65a8a9acc6147457db8782b293e49770f38b81
Change-Id: I62aa88001d2d7b2fdaff9c97cd406e08d4d81949