metrics(AddImage): record cache hit rate too
Currently, we only record the cache miss rate when called via ::get
which happens in the BeforePageDisplay hook. This should be close to 0.

However, we can not interpret this data, because we have no number for
the cache hits to compare it to.

TODO: find a way to actually test this locally to assert that it works
as expected

Bug: T387877
Change-Id: I113f9b8d81c1812cd64300668f43752bf3405135