fix(CC2.0): enforce max-limit on InfoboxTemplates
This was previously enforced in the legacy on-wiki-config in
GrowthConfigValidation. There it was set to 800 in change I2b9ec69a3056
(commit 3a8d938).

Counting all the existing configurations for that field in all the
wikis showed that non exceeded even just 600, with only Polish Wikipedia
exceeding 500. Therefore, we can move forward with adding the limit here
without requiring a migration or additional mitigations or workarounds.

We have to inline the `[ 'class' => MediaWikiDefinitions::class, 'field'
=> 'PageTitles' ]` template, because setting a limit on it "from the
outside" doesn't work. It would be nice to rework that functionality,
but that probably exceeds the scope of this task.

With respect to the upcoming removal of CC1.0 code, including
GrowthConfigValidation (T388227), this change also already replaces
usage of that constant in our code-base.

Bug: T386629
Change-Id: Ibfde84f858ebfc540a45aba23a5fa358ea473111