feat(CommunityUpdates): set a white background in night mode based on config
Add a new boolean property in the config schema to
support the setting. Then render an extra css class
--dark-mode for the image setting a white fixed
background. The class is prefixed by .client-darkmode
and .skin-theme-clientpref-night and clientpref-os
to only take effect on these modes.

Also change the the container for the background image
from span to a div since the inline display is preventing
the image to show on first render.

Bug: T387737
Change-Id: I4053b901fa0a42a0d86b68354e3941b621d9ad16