fix(NewcomerTask): realign pulsating dot on edit
This broke likely when Codex-styles were introduced to Minerva.

After the introduction of codex-styles, neither the link itself nor its
`<li>` had `position:relative` anymore, and that caused the position of
the blue dot to sometimes be way off. Fix that by adding
position:relative to the `<li>`. We cannot add it to the
link itself, because then the bouncing effect would be cut off at the
lower edge.

Also, for narrow screens that button is supposed to be icon-only.
However, the codex styles to hide the label are matching on
`span:last-child` and that doesn't work with us appending the blue dot.
Prepend it instead.

Further, the media-query breakpoint for showing the button labels is now
1119px, so we follow that.

Finally, likely with the switch to Codex as well, the blue dot has
become vertically centered in the icon, which does not correspond to the
original designs from T244435. So this change now explicitly moves it
further down to the lower edge, realigning it with the designs again.

Bug: T388602
Follow-Up-To: I7bf3690f1ab9c9e4dd71761a03fa22e19b712c67
Change-Id: I73a327afe752e9441810812be23408d6a0aa94b9