Always return a function score builder for growth_underlinked
If we don't return a function score builder here a request
invoking this will result in an Exception bubbling up to the top.
That generally isn't a desirable method for handling errors,
particularly because the api will claim that this is a supported
rescore profile. Instead return a builder that does nothing.

Bug: T388898
Change-Id: I8a6986319cb6e9cd9bc85b49e97785165f88411e