homepage: rename _duration to _loadeventend
It is very unusual to access navigationEntry.loadEventEnd as
'duration'. This is not in any way the name of this metric, as it
exists for interop with type-agnostic consumption (everything has a
duration). The two are identical per [1].

What you want is loadEventEnd, which is what this is called everywhere
else, and will never change in meaning. It is also how it is collected
elsewhere in MediaWiki, Chrome's CruX report, and other performance
measurement software.

Given that this metric was only added to Prometheus last week,
now seems like a good time to fix the metric name before it gets
baked into long-term dataset continuity.

While at it:

* The name also seemed to lack any mention of "homepage" which I
  assume was not intentional. I see this was mentioned at [2] but
  I suspect it was not realized that labels cannot be added retroactively
  without breaking continuity.

* Fix paint access to retreive by type first, since names are
  local to theire type (not globally unique). This should be a no-op
  but similarly matches how the data is retreived elsewhere by
  MediaWiki, by Google, etc. and reduced doubt in the data going
  forward.

[1]: https://www.w3.org/TR/navigation-timing-2/#performanceentry
[2]: https://gerrit.wikimedia.org/r/c/mediawiki/extensions/GrowthExperiments/+/1109391/17/modules/homepage.init.js

Bug: T382003
Change-Id: I5f95901b9d0bc69bc38a66a99c7bb67c6a10f4f3