fix(surfacing): don't surface on old revisions or diffs
Apparently, viewing diffs and old revisions still has the `action`
`'view'`. So we need a different way to check for these circumstances,
and `oldid` seems to be the way to detect viewing a specific (old)
revision, whereas `diff` shows up when viewing diffs. Both can stand on
their own, so we do need to check for both.

So only show the surfaced recommendations when these attribute do not
exist in the URL.

Change-Id: I38febfaed8a1019a74bbca7b15fec434918d9039