CText: Replace custom type scale with Codex tokens
Once Codex "font modes" are released, the custom type
scale used in the CText component here can be replaced
in favor of using the appropriate line-height and font-size
tokens from Codex. This change should not be merged until
Codex 2.0 (which will include the font modes feature) has
been released, because some of the tokens used here are not
available yet.

One additional benefit of this patch is that it will allow
CText components to automatically adapt to any font mode
switching that happens based on skins while preserving the
visual hierarchy between elements.

This patch is set to depend on the relevant font modes patch
upstream in Codex, but really it should be marked as depending
on a yet-to-be-created patch that adds Codex 2.0 to MW core.

Bug: T391018
Depends-On: I700c27fb437d253eec0c399546681c50af7b3cb3
Depends-On: I0ec2198617118272cc3c3e97d9e592de910cb7b5
Change-Id: I7d96d57f14fb6e76e0e8e70b07a629053cad334a