chore: update stylelint and disable broken rule
Unfortunately, in 0.18.0 the `declaration-no-important` is broken and
needs to be disabled.

This change also disables the report-needless-disables rule, because
those comments still signify which usages of `!important` have been
intentional and well-considered.

Hopefully, the upstream bug will be fixed soon and this commit can then
be reverted. Until then, this is one piece to unbreak LibUp for
GrowthExperiments.

Bug: T387519
Change-Id: I9c1e9f15d2650d037c6e5e03c9fdadf8126833fb