chore: update eslint config
Disable unicorn/prefer-includes to make allow the lint task
to pass. `.indexOf` instances are replaced in the next change.

The code in `documentation/frontend/` was actually always linted twice
until now, but now this breaks because when it is being linted from the
root directory, then its npm dependencies are not yet installed and it
fails with a no missing import violation,
Now we're only linting that directory in its dedicated `npm run test`
call. This might also very slightly reduce the overall runtime of this
job. Though we should really look into actually making use of that code
and introducing it back into our main directory structure.

Change-Id: Ic610be785a7ff89870165857cc71883d4d6ecf7b