perf(Mentorship): batch filtering mentees in UncachedMenteeOverviewDataProvider
This should make sure that the work of any individual database query is
predictable as a mentor can have many hundreds of mentees.

The batch-size can be set via the command-line option of the maintenance
script.

The default is lowered to 100 per the recommendation from the task.

Bug: T391695
Change-Id: I14451a2b3cee88ad7dcb13da41382228bc70a30f
(cherry picked from commit c7a52adb66fd48b25388890e48179f72f4012f49)