cleanup(GENewcomerTasksTopicType): sunset topic types
GE worked in the past with "growth" and "morelike" topic types but
that's no longer in use in any wiki and there's almost no value
on keeping the topic type configurable.

This change is to remove the config var and cleanup the associated
user preference growthexperiments-homepage-se-topic-filters in
favor of growthexperiments-homepage-se-ores-topic-filters.

Bug: T390167
Change-Id: I48fad2711ff505436716b530590b70b75fda02a6