fix(PostEditPanel): re-render main area when task queue finishes loading
When the daily quota for a task is reached, the results for 'growthnextsuggestedtask'
request for the exceeded task type are filtered in the client. In
Ib545b9f24ea6b84504a8ba4d2c4561737be9be0b the order of events when the
task queue is updated and finished loading where switch. The post edit
dialog was only re-rendering after the task queue was updated but
the loading flag was not yet updated and propagated.

Add a TASK_QUEUE_LOADING listener to re-render the dialog once the
tasks loading has finished.

Bug: T383433
Change-Id: I59d5b1686754b2c64df89a587753a46fd7cc40ae