chore: factor eslint and stylelint out of Grunt
This gets us closer to a conventional modern setup of our frontend
linting and a step closer to ditching Grunt altogether. Now we depend a
bit less on poorly maintained Grunt plugins. We continue to use Grunt
for Wikimedia's homegrown banana-checker, because it seems to basically
exist only as Grunt plugin.

A follow-up change should introduce npm-run-all, because the `npm test`
command is getting large.

Change-Id: If5da4a55f81683294b863dabf46aa46af1db1fca