metrics(refreshLinkRecommendations): only record integer with a counter
The `refreshLinks_session_seconds_total` metric is not producing the
expected data and I do not know why that is. The code looks plausible
and I cannot see any obvious bugs. While I doubt that it is the cause of
these issues let's try increasing the counter only by integers, maybe it
helps.

Change-Id: I2a423723d4bd73f893eb38e42bd66bc2e7ec0cd6