SpecialHomepageLogger: Populate email state even with StartModule disabled
start_email_state is a required field on the schema. However, we do
not populate it if StartEmail module is disabled. Previously, it used
to be always enabled, but T275155 changed that.

This patch computes the state in logger itself to satisfy requirement.
We can consider making it non-required or removing it altogether
if needed, but probably later.

Bug: T394017
Change-Id: I2f9183b9c19372c54b629a19e14bff176a628926