feat(AddLink): show task as unavailable when user editcount reaches max
- Add TaskTypeManager to centralize config and user aware logic
for task types, such as limiting the AddLink task based on edit count
- The manager will add a suggested next task to the user task types
filter if the result of filtering unavailable tasks (tasks that should
not be selectable although visible) is and empty filter. The logic
to suggest a new task is provided by LevelingUpManager
- Update newcomerUserOptionsLookup::getTaskTypeFilter callers to use
TaskTypeManager::getTaskTypes
- Adapt LevelingUpManager::suggestNewTaskTypeForUser to use a list
of available tasks as input instead of directly getting the task
types from config. This also removes the need for the duplicated
A/B link/link-recommendation handling in getTaskTypesGroupedByDifficulty
- Add a new property 'unavailableTasks' in suggested edits module
action data 'wgGEHomepageModuleActionData-suggested-edits' to ease
the frontend disabling of unavailable tasks.

Bug: T393769
Change-Id: Ib9c2e7e6dd492d5dfababbb5739ebf32a4bb2b29