LinkRecommendationStore: Query templatelinks on the main DB
Previously, LinkRecommendationStore was querying for templatelinks
on x1, where it does not exist. This patch fixes that by injecting
the connection provider and querying for the correct database.

At some point, we should migrate the class to IConnectionProvider
fully (together with rest of GE). This patch does not do that yet,
as it is non-trivial and it aims to fix a train blocker.

Bug: T396680
Change-Id: Id64d89f0ac2f61236e5e500155ed76c5c7702dbb