refactor: Schedule LevelingUp notifications from LevelingUpManager
Why:
With T394957, the code for scheduling LevelingUp notifications will
soon get slightly more complex (by adding variants into the play).

To be able to handle such complexity, let's move the scheduling
code to LevelingUpManager, where it can be centralized.

Conceptually, it might be better to have this code in LevelingUpHooks
instead, but that is tricky due to call order (we do not know what hook
handler code runs first).

What:
Factor the notification firing code to LevelingUpManager

Bug: T394957
Change-Id: I000c23ab7a90d86fa04e6bd65dc861a917e55f3d