feat(LevelingUp): Measure the delay between actual and intended notification timestamp
Why:
Delayed notifications are handled via delayed jobs. By design,
those jobs never run at the exactly scheduled timestamp, but slightly
before/after it. As the Growth team, we have no visibility into
how precise the mechanism is.

In T392256, the Growth team wishes to schedule the notification
differently for different users. Based on an engineering
spike (T393955), we are not sure how reliable that change
would be.

To ensure we are within the tolerated error margin (1 hour),
we need to measure the accuracy of the notification delay.

What:
* Introduce a common layer between NotificationGetStartedJob
  and NotificationKeepGoingJob, called AbstractDelayedNotificationJob
* Add measurement code as AbstractDelayedNotificationJob::measureNotificationDelay.
  This uses GaugeMetric, because the metric value can be both positive and negative,
  which histogram doesn't support.
* Add public interface for getting the configured notification delay
  to LevelingUpManager and cover it via tests

Note:
This seems like potentially upstreamable code, as it might be useful
to other delayed jobs as well, not just us.

Bug: T395260
Change-Id: I625ad95f87be61d55f9a4c76111990c892db76c2
(cherry picked from commit 1c0872b21fd7adbe050ef436f8fda9403b0076bd)