StructuredTaskMobileArticleTarget: Fix definition of toolbarGroups
This will prevent the tools being out of order when upstream
cleans up mobile toolbarGroup hacks (Id9d587b3d73).

Ideally we want to copy the definitions from upstream by name
but we have a circular dependency with Id9d587b3d73 so that
can happen in a later commit.

Change-Id: I560ae6ff1384890fbd5ffeca5a70f333384d4fe5