fix(Mentorship): Prevent too long mentor message
In StructuredMentorListValidator::validateMentorMessage we check that
the message is not longer than the limit. However, if that is the case
we only add a warning, not an error, to make sure that mentorship still
works as a whole. However, we did not consistently check for warnings
when saving the list of Mentors as we should have and that allowed the
too long message to be saved via editing the json directly (where we did
not check), while preventing any modification via the UI where we do
check that the entire list is valid.

Bug: T399086
Bug: T384445
Change-Id: I6fffd6f482870243e0f910bbefdee2d523f6bf30