fix: undo mistakenly changed notification copy
This copy was changed in I7575eee0b (commit 9964b35), but it was
intended to only affect the new celebratory notification when reaching
the community-defined Add a Link milestone. However, since we reused a
message for the CTA instead of creating a new one, changing the i18n
file also affected the Get Stated notification.

This change undoes the change to the i18n file and introduces a new
message for the CTA of the new notification.

Bug: T398262
Change-Id: I5bdf8e0afed19e86fb7fc254dad499a9514798b7