Replace server-side duration tracking with Performance API
This change addresses a timing issue where we were
calculating the server-to-client duration by combining
server and client timestamps. This approach was
problematic because it sometimes resulted in
negative duration values due to clock
synchronization issues between the server and
client machines. Instead of using
`mw.now() - mw.config.get('GEHomepageStartTime')`,
this patch provides a  measurement from when the
request started on the client until the Suggested
Edits module is fully initialized.

Bug: T395627
Change-Id: I1df451c8a5f1e5fadc56e12ff752f7e480d367f2