chore: update eslint-config-wikimedia to unbreak LibUp
With version 0.31.0, eslint-config-wikimedia now includes the typescript
linting-configuration that we formerly got from
@wmde/eslint-config-wikimedia-typescript. Also, these rules are now
automatically applied to all *.ts files.

This should now hopefully unbreak LibUp, which was so far blocked on
version conflicts between @wmde/eslint-config-wikimedia-typescript and
eslint-config-wikimedia.

Also, this adds --max-warnings=0 because we currently have no eslint
warnings and I think we should try to keep it that way.

Change-Id: I19faedb805502dce731b32fd7392d065d618a05a