LevelingUpManager: show different notification based on experiment groups
Introduce experiment variant checks where appropriate in
addition to the existing feature flag checks for
GELevelingUpNewNotificationsEnabled.

Pass a new param eventType to the GetStarted and KeepGoing
notification jobs to show different presentation model
based on control/treatment variants. Control should
get the existing copy and treatment the "new".

For the ReEngage notification, which is newly introduced
only for treatment group, only schedule it if the feature
flag is enabled and the user is assigned to the treatment
variant to prevent creating a no-op job.

Bug: T401308
Change-Id: I4d4f89c940d13b902be1a9a546ba4ef8e87cda44