SECURITY: Ensure the mentor is registered before waiving permission
Previously, ApiSetMentor allowed anyone to pretend they are a mentor
and claim mentees. This was not the intention; instead, we wanted to
allow claiming for _registered_ mentors (the equivalent of Special:ClaimMentee,
but in the API).

This patch allows the change of mentors to happen only if one of those
conditions is met:

  1. The target mentor is registered, and the performer is either
     the mentee or the mentor.
  2. The performer has the `setmentor` permission (normally
     available to sysops)

In addition to the Phabricator-reviewed version, this patch skips
the test that is (intentionally) broken by this patch.
I607e4cf01b0b6f5a1b84ddb5f220adc12bcbfecd is a (CI-passing) follow-up
that fixes this, and that will be reviewed through Gerrit.

Bug: T402600
Change-Id: I29a18dbbaf7e2ce2a713233dbc6880032fec3628
(cherry picked from commit ac5d60083b10314ce1a1c60b7668e7ed275b5248)