fix(QuestionPoster): Use getPermissionStatus
QuestionPoster was previously using getPermissionErrors(),
which returns errors as an array. This array was then converted
into a Status, which -- errorneously -- contained HTML instead
of message keys.

PermissionManager also provides getPermissionStatus(), which returns
a PermissionStatus. This is much easier to convert to a Status, so it
seems more appropriate to use for our use case.

In addition to that, getPermissionErrors() was deprecated, so this
is an useful change even without the HTML bit.

This patch changes the signature of QuestionPoster::checkUserPermissions
to return a StatusValue (prior to this patch, it returned a Status).
This is done, because PermissionStatus inherits from StatusValue, but
not from Status. We can promote a StatusValue to a Status by calling
Status::wrap(), but since we actually do NOT need a full Status, there
is no point in doing so. In our tests, we already mock checkUserPermissions
so that it returns a StatusValue anyway :).

Bug: T401096
Change-Id: I5caa5a457308a5e96dc0969e16635264d2657e5d