Hide Suggested Edits link when SE module is disabled
On wikis where the Homepage is enabled but the
Suggested Edits module is not, Special:Contribute
and Special:Contributions pages currently displays
a link inviting users to "Suggested edits" or to
"Visit your homepage to get started with editing".
Clicking the link takes users to Special:Homepage,
but without the SE module there is no editing guidance,
which can be misleading.
This change checks the SuggestedEdits config before
rendering the link, and omits it entirely when the
SE module is disabled.

Bug: T403558
Change-Id: I16db1cb78da17d070b161b22fd175c53b3c7279d