LevelingUp: setup AB test for notifications
Make use of two existing flags to start collecting
experiment related data:
 - wgGENotificationsTrackingEnabled enables the load
of the client module for notifications instrumentation
(both long-lived and experiment related).
 - wgGELevelingUpNewNotificationsEnabled is the feature
flag for the new notification texts and audiences to
kick-in. It should be enabled in the target wikis
for the experiment configured via mpic dashboard.

Do not collect the long-lived metric instrument data
in wikis with the notification experiment enabled. That
is to prevent the treatment experiment group affecting
the ling-lived baseline metric.

Bug: T401308
Change-Id: I01782d7b0bcddd3542aeec85794f6aedc7616e47