test(HelpPanelProcessDialog): assert correct preferredEditor is set
In the precedent change we spotted a wrong type for a qunit test
mock, but since that had not yield any failure in years, it was
probably not being read at all. Update the mock and test code
to make it assert the preferred editor pick up for two
different tasks.

Follow-up: I2d7cd9818e9b7c603575a3692ae8749f98987c1b
Change-Id: I72db4f825433a320d75e1d463c4a2219e8e9d23d