Remove meaningless fill-rule attribute from SVG files
This re-applies parts of the reverted patch I7a28ccd. But much
slower, one step at a time.

fill-rule can only have an effect on shapes with holes, for example
the robot's eyes. But even then it's only needed when the outer and
inner shapes are drawn in opposite directions. This is exceptionally
rare. In this codebase only 4 fill-rule are actually needed.

I manually checked each file touched in this patch and verified that
they really do not change.

Change-Id: If22ac9d08c39c2cfa3f0c25b9257a9bd0fe98f45