Remove meaningless clip-rule attribute from SVG files
This re-applies parts of the reverted patch I7a28ccd. But much
slower, one step at a time.

From MDN: "The clip-rule property only applies to SVG elements that
are contained within a <clipPath> element".
https://developer.mozilla.org/en-US/docs/Web/CSS/clip-rule

While some of the images in this patch do use clipPath, they never
use the clip-rule as part of the clipPath.

I also manually checked each file touched in this patch and verified
that they really do not change.

Change-Id: Ie6fec9d18e589959234471a9341fce25df2032b6