Remove unused id="…" attributes from SVG files
This re-applies parts of the reverted patch I7a28ccd. But much
slower, one step at a time.

An id is only meaningful when the element is somehow reused. This
is the case when e.g. a filter is used via filter="url(#a)" or an
element reused via xlink:href="#a". This is not the case in the
files touched in this patch.

I manually checked each file touched in this patch and verified that
they really do not change.

Change-Id: I320bd4cbb873e09b489c8bd5367a45e096e7ca58