feat(ReviseTone): Show Tone Check
This change now brings in the integration with the work happening on the
VisualEditor side. Unfortunately, this still means that there is a
rather tight integration with VisualEditor that is very hard to
unit-test.

Changing the name of the Tone EditCheck has a couple of non-obvious
consequences:
  * we need to listen to different events to detect the user declining
	the tone check in Ia02fe5854af796ae7b41297adb63242d78f5d00d
  * the on-wiki config for the ToneCheck does not apply to our check,
	but that is probably desirable
  * this will affect the events logged to analytics: on the one hand,
	this will make it easier to distinguish our check from Tone Check
    checks in the data. On the other hand, an inclusive analysis would
    need to know about both.

On mobile, there is still the issue that the tone check starts out as
minimized. This will either be fixed as part of T404666 or as part of a
yet-to-be-created task.

Bug: T401195
Change-Id: If4dcda23375a94a1721a46c91c284162c582a3e0