Skip tests requiring CirrusSearch when it is not loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When GrowthExperiments is a dependency, we are required to
add CirrusSearch to have tests to pass.

By making CirrusSearch optional, on CI that saves us from having to add
CirrusSearch and its dependencies as dependencies of extensions
depending on GrowthExperiments.

Turn some unit tests to integration tests which is required in order to
use `markTestSkippedIfExtensionNotLoaded('CirrusSearch')`.

Move some unit/integration tests common code to a trait
HomepageHooksHelpers.

Tested using: phpunit extensions/GrowthExperiments with the two
following configurations:

- With CirrusSearch (and its dependency Elastica):

    wfLoadExtensions( [
        'GrowthExperiments',
        'CommunityConfiguration',
        'VisualEditor',
        'CirrusSearch',
        'Elastica',
    ] );

- Without CirrusSearch:

    wfLoadExtensions( [
        'GrowthExperiments',
        'CommunityConfiguration',
        'VisualEditor',
    ] );

Bug: T415278
Change-Id: Ib7ca8a72c8e0f16aff3d8350a21d50f1aa0000cb