Make image suggestion feedback optional
On CI, we would like to remove the recursive injection of extensions
dependencies. When GrowthExperiments is a dependency, we are required to
add EventBus to fullfil the services instantiations.  Without EventBus,
MediaWikiServicesTest::testDefaultServiceInstantiation structure test
fails with:

  1) MediaWikiServicesTest::testDefaultServiceInstantiation
  Wikimedia\Services\NoSuchServiceException: No such service: EventBus.EventBusFactory
  ...
  extensions/GrowthExperiments/ServiceWiring.php:202
  ...

This is caused by the ImageSuggestionFeedbackUpdater requiring EventBus.
By making it optional, on CI that saves us from having to add EventBus
and its dependencies as dependencies of extensions depending on
GrowthExperiments.

A previous hard dependency on EventBus has previously been fixed by
Ic619318a1d74aaf5b34eaab1346231642bc03eba (T329672).

Bug: T389998
Bug: T329672
Change-Id: I51cb33d8836dd97c397b218969d33cfdd7458bef