maintenance(purgeExpiredMentorStatus): dont bail out on wikis with invalid config
Somewikis have deleted the config file MediaWiki:GrowthMentors.json
that was created during the initial migration from CC1 to CC2. Some
were already locked at the time and did not get the file. While
it would also be appropriate to fully remove GE from these wikis
that wouldn't prevent the failure to happen again for wikis that
become locked or accidental deletions of the config file.

Check for an initial valid config before attempting to purge it
and just output the invalid error when it exactly matches
growthexperiments-mentor-list-missing-key which is the error
thrown by the mentor list validator when there is no top
level config key. Which is what happens when trying to load
a wiki page config for an inexisting page.

Bug: T407366
Change-Id: I7b778895627b8d1d8991fa922250a6b4578ef4f6