tests(Cypress): Skip second Revise Tone mobile browser test
This test was supposed to be fixed with
Id0a955bac8625daab0b07c82224783700e0a2c32, but evidently this does not
work as it is failing with the same error now in Gate-and-Submit. This
needs more investigation. Let's skip that test too for now. This means
that we now have no browser tests for mobile, only for desktop.

Bug: T407152
Bug: T407715
Change-Id: Ice9ada03b46cbd87dddcef7ac19dbc583244c040