feat(ReviseTone): add production search terms
During the duration of the A/B test, we're not making use of
CommunityConfiguration yet. The reason is that we want to have more
control over the task during the duration of the test, and don't want to
introduce configuration that we then may have to remove again later.
So some essential configuration will be provided by PHP settings. Once
we decided that we want to keep the task permanently, much of these
configurations will move to community configuration.

Bug: T407353
Change-Id: I10b3945f8e88d5a8105fe9114ed6a88226a5ad54