ReviseTone: A/B test experiment setup
Check group assigned by xLab to determine if RT recommendations
are enabled for a user. The check is only performed when
wgGEUseMetricsPlatformExtension is enabled to allow wikis
without on-going experiment to show the task based only
on the feature flag.

Callers willing to get the filtered list of task types for a
user should call NewcomerTasksUserOptionsLookup::getTaskTypeFilter

Bug: T402707
Bug: T407528
Change-Id: Icec36a3902fa432806db84b893a1af1de5900ec7