fix: don't run listTaskCounts if Newcomer Task are not available
The errors in T408052 are downstream from T369312. Similar issues are
being fixed in T408531.

This script should not be run on wikis which have Suggested Edits
disabled (currently, frwikitionary, in the future maybe Wikidata). Aside
from adjusting the job runner that executes this script, exiting early
is also a good idea and that is what this change implements.

Bug: T408052
Bug: T408531
Change-Id: I98d7584960ce97417a1d9344a19959d0693d4144