feat(ReviseTone): finish i18n work of onboarding quiz
This keeps the message keys for the quiz itself in the quizData.json
file, because we need to have them somewhere for findability purposes
and I rather not have 30 lines of message comments in the main
onboarding component.

The link is kept in the description message, because that makes it more
likely that they are translated consistently.

The `growthexperiments-revisetone-onboarding-dialog-title-label` message
has been split into two because half of it changes for every step as
well.

Bug: T407580
Change-Id: I9ca9367f8f9f89f66cf2b5a89664792a86dd857c