feat(ReviseTone): drop weighted tag on missing recommendation data
When there is no recommendation for a revision in Cassandra, then that
is probably due to a race-condition on a very recent edit. (That is to
say, sometimes it takes CirrusSearch a few minutes it learns that a page
was edited. After CirrusSearch learned that, it is automatically
excluded by our query on production.)

Another potential cause for this might be an issue with the ML update
pipeline which usually should take care of adding and clearing the
weighted tags.

Bug: T407538
Change-Id: I97a3b0a1235731c7558b8399eb0a779d88338afb