fix(ReviseTone): do not flash other task on homepage
Previously, the Revise Tone task was only put at the front in the js
data. While that worked, it meant that another task was shown while that
data was still loading. This change fixes that behavior by sorting the
Revise Tone task to the front with a filter with happens earlier and
affects all the places where that list of tasks is shown on the
homepage.

Change-Id: Ib3b05414803d07e9805562c1c681690fd434703c