fix(ReviseTone): restore modal when task in not available
This fixes the error of the modal not showing for Revise Tone because
the code tries to log an event to analytics and that crashed because the
logger was null. Now with this change the logger is set to a very basic
implementation. Further refinement will happen in context of T406252.

I'm unsure whether `action_subtype` is right field, but for now this
would just log `action=impression` and
`action_subtype=nosuggestions_dialog`

Bug: T407538
Change-Id: I0b435e48361559cf71af5204ae4e24cddcaf21ee