fix(ReviseTone): only initialize once
It turns out that the CodeMirror extension with the configuration
`$wgCodeMirrorV6 = true;` is somehow triggering the `ve.loadModules`
hook a second time, and somehow that seems to cause the addPlugin
callback, and thus the initializer code, to be run 3 additional times.

In terms of Wikipedia, this configuration only exists on testwiki and
Hebrew Wikipedia. See T357795.

To work around that issue, this change introduces a static variable on
the ReviseToneInitializer class to check if it already has been
initialized and exits early in that case.

Change-Id: Ib5d0a10244d003fef5b2eadbc13af466859dcb27