Handle exception when mentee asks a question to a nonexistent mentor
Previously, if a user asked a question to a mentor in a wiki that has no mentors, the code would
throw an unhandled exception. After this change, the UI instead displays an error message to the
user. Users will also see an error message if the asks a question to their nonexistent mentor after
opting out of mentorship.

Bug: T386567
Change-Id: I102554c9e55eae9917f527daa6d312170cb3bce0