fix(ReviseTone): render behind EditNotice on mobile
This fixes an issue that the onboarding was not interactable on testwiki
on mobile, because an edit notice had opened behind it.

Both the edit notice and the onboarding's vue dialog are teleported to
the same container. In particular, the onboarding is showing while the
editor is still loading, and so it is teleported first and is showing up
first in the DOM order in that element. The edit notice starts to show
up once the mobile editor has finished loading.
However, both the edit notice and the onboarding have a z-index set. The
onboarding has it set to 400 (@z-index-overlay-backdrop), but the
backdrop of the edit-notice has a z-index of 4 (four). But somehow,
these two together result in the strange experience that the edit-notice
itself is behind the onboarding dialog, but its backdrop still is
disabling interacting with the onboarding dialog.

This change resolves that issue by not teleporting the dialog in the
first place. The means that the edit notice is showing in front of it
and can properly interacted with. From early testing, I have not noticed
drawbacks from not teleporting the dialog. It should be fine because we
are appending the mount-point to the body tag.

Change-Id: I85f21ce7fe68b1e6b4ba5071428a01e48e168971