fix(ReviseTone): don't show Revise Tone Check twice
Recent changes to VisualEditor, including I5c0c4f9b seem to have caused
the issues we're seeing with Revise Tone showing up multiple times. In
particular, the change mentioned above changed the `setup` method in the
`controller.js` file to no longer call `this.onDocumentChange()`, but
instead it calls now `this.refresh()`. This now executes all mid-edit
listeners, not just `onDocumentChange`, but also `onBranchNodeChange`.

To achieve that consistently, we must make that code path trigger even
after the Editor has already finished loading. This can be achieved by
re-emitting the 'surfaceReady' event on the editor target. However, in
the case that the onboarding has already been permanently closed, we
must make sure we do not fire the 'surfaceReady' event prematurely. We
can in that circumstance just rely on our edit check show up in course
of the usual event fired by VE.

That means that triggering the EditCheck via onDocumentChange
and 'undoStackChange' is no longer necessary, and now even counter
productive because it leads to two of our ReviseTone Edit Checks showing
at the same time, one from onDocumentChange and one from
onBranchNodeChange.

Note that part of the code that we are deleting here was originally
added only October 7th in Ic58eeac03f and in particular in this comment:
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/GrowthExperiments/+/1198088/comment/db3cb4bf_fee863ca/

The code triggering the stack undo was originally introduced in
If4dcda23375a94a1721a46c91c284162c582a3e0

In general, it is a very bad sign that crucial code that we're relying
on is changing without us noticing or a test failing. This is likely
going to be very brittle going forward. That makes it crucial that we
fix and reenable the browser tests as soon as possible, even just in
limited scope that only asserts that the check shows up exactly once and
does not interact with it or the text.

Change-Id: I8af975356ed27ab7be04b01d1f7e036d8ff4df85